lint

+++ contracts/contracts/utils/DepositContractUtils.sol
@@ -2,28 +2,35 @@
 pragma solidity ^0.8.0;
 
 contract DepositContractUtils {
-    
     function calculateDepositDataRoot(
         bytes calldata pubkey,
         bytes calldata withdrawal_credentials,
         bytes calldata signature
-    ) public pure returns(bytes32 node){
-        uint deposit_amount = 32 ether / 1 gwei;
+    ) public pure returns (bytes32 node) {
+        uint256 deposit_amount = 32 ether / 1 gwei;
         bytes memory amount = to_little_endian_64(uint64(deposit_amount));
 
         // Compute deposit data root (`DepositData` hash tree root)
         bytes32 pubkey_root = sha256(abi.encodePacked(pubkey, bytes16(0)));
-        bytes32 signature_root = sha256(abi.encodePacked(
-            sha256(abi.encodePacked(signature[:64])),
-            sha256(abi.encodePacked(signature[64:], bytes32(0)))
-        ));
-        node = sha256(abi.encodePacked(
-            sha256(abi.encodePacked(pubkey_root, withdrawal_credentials)),
-            sha256(abi.encodePacked(amount, bytes24(0), signature_root))
-        ));
+        bytes32 signature_root = sha256(
+            abi.encodePacked(
+                sha256(abi.encodePacked(signature[:64])),
+                sha256(abi.encodePacked(signature[64:], bytes32(0)))
+            )
+        );
+        node = sha256(
+            abi.encodePacked(
+                sha256(abi.encodePacked(pubkey_root, withdrawal_credentials)),
+                sha256(abi.encodePacked(amount, bytes24(0), signature_root))
+            )
+        );
     }
 
-    function to_little_endian_64(uint64 value) internal pure returns (bytes memory ret) {
+    function to_little_endian_64(uint64 value)
+        internal
+        pure
+        returns (bytes memory ret)
+    {
         ret = new bytes(8);
         bytes8 bytesValue = bytes8(value);
         // Byteswapping during copying to bytes.
@@ -36,5 +43,4 @@ contract DepositContractUtils {
         ret[6] = bytesValue[1];
         ret[7] = bytesValue[0];
     }
-    
 }
